requirements.txt
asgiref==3.3.1
Django==3.1.4
django-markdown-deux==1.0.5
django-pagedown==2.2.0
markdown2==2.3.10
Pillow==8.0.1
pytz==2020.4
sqlparse==0.4.1
